pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
python-dotenv = "^0.19.0"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.25"
aiosqlite = "^0.17.0"
asyncio = "^3.4.3"
asyncpg = "^0.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
